add new paginate totaldaily

+++ internal/domain/entity/summary.go
@@ -4,13 +4,13 @@ import "time"
 
 type SummaryDashboard struct {
 	ID             int64     `gorm:"primaryKey" json:"id"`
-	TotalActiveSub int       `gorm:"size:10:default:0" json:"total_active_sub"`
+	TotalActiveSub int64     `gorm:"size:10:default:0" json:"total_active_sub"`
 	TotalRevenue   float64   `gorm:"size:15:default:0" json:"total_revenue"`
 	CreatedAt      time.Time `json:"created_at"`
 	UpdatedAt      time.Time `json:"updated_at"`
 }
 
-func (e *SummaryDashboard) SetTotalActiveSub(v int) {
+func (e *SummaryDashboard) SetTotalActiveSub(v int64) {
 	e.TotalActiveSub = v
 }
 
@@ -26,7 +26,7 @@ type SummaryRevenue struct {
 	ID        int64     `gorm:"primaryKey" json:"id"`
 	Subject   string    `gorm:"size:45" json:"subject"`
 	Status    string    `gorm:"size:45" json:"status"`
-	Total     int       `gorm:"size:10:default:0" json:"total"`
+	Total     int64     `gorm:"size:10:default:0" json:"total"`
 	Revenue   float64   `gorm:"size:15:default:0" json:"revenue"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
@@ -34,9 +34,9 @@ type SummaryRevenue struct {
 
 type SummaryTotalDaily struct {
 	ID           int64     `gorm:"primaryKey" json:"id"`
-	TotalSub     int       `gorm:"size:10:default:0" json:"total_sub"`
-	TotalUnsub   int       `gorm:"size:10:default:0" json:"total_unsub"`
-	TotalRenewal int       `gorm:"size:10:default:0" json:"total_renewal"`
+	TotalSub     int64     `gorm:"size:10:default:0" json:"total_sub"`
+	TotalUnsub   int64     `gorm:"size:10:default:0" json:"total_unsub"`
+	TotalRenewal int64     `gorm:"size:10:default:0" json:"total_renewal"`
 	TotalRevenue float64   `gorm:"size:15:default:0" json:"total_revenue"`
 	CreatedAt    time.Time `json:"created_at"`
 	UpdatedAt    time.Time `json:"updated_at"`
